two working directives
