+++ app/src/components/magazines-list/magazines-list.html
@@ -1,0 +1,35 @@
+
+<div>
+    <h3>My News Resources</h3>
+    <p>
+        name: <input ng-model='search.name'><br>
+        min age: <input ng-model='search2.age'><br>
+         {{ search | json }}
+    </p>
+    <table>
+        <thead>
+        <td>number</td>
+        <td>name</td>
+        <td>age</td>
+        <td>comment</td>
+        <td>url</td>
+        <td>percent change</td>
+        </thead>
+
+        <tr ng-repeat="magazine in magazines
+			| filter: search
+			| filter: search2
+			| orderBy: 'age'
+			track by magazine.name ">
+
+            <td>{{ $index + 1 }}</td>
+            <td>{{ magazine.name }}</td>
+            <td>{{ magazine.age }}</td>
+            <td>{{ magazine.comment }}</td>
+            <td>{{ magazine.url }}</td>
+            <td>{{ magazine.change }}</td>
+
+        </tr>
+    </table>
+
+</div>
