+++ app/src/data/data.js
@@ -1,0 +1,20 @@
+import angular from 'angular';
+
+export default dataarray;
+
+var dataarray= [
+
+    { name: 'New York Times',   age: 24, comment: 'centrist', url:'http://google.com', change:'23' },
+    { name: 'Washington Post',  age: 8 , comment: 'right wing', url:'http://google.com',change:'54' },
+    { name: 'Huffington Post',  age: 4 , comment: 'semi-serious', url:'http://google.com',change:"43" },
+    { name: 'Daily Beast',      age: 10 , comment: 'newmedia', url:"http://google.com",change:"28"},
+    { name: 'Atlantic',         age: 14 , comment: 'respectable',url:"http://google.com", change:"27"},
+    { name: 'Wall Street J',    age: 33 , comment: 'capitalist', url:"http://google.com",change:"26"},
+    { name: 'NY Post',          age: 22 , comment: 'street rag', url:"http://google.com",change:"25"},
+    { name: 'LA Times',         age: 23 , comment: 'once great', url:"http://google.com",change:"24"}
+
+];
+
+
+
+var dataarray = module.exports(dataarray);
