+++ app/src/filters/index.js
@@ -1,5 +1,17 @@
+import angular from 'angular';
 import momentFilter from './moment-filter';
 
-export default function( ngModule ) {
-    momentFilter( ngModule );
-}
+const filters = angular.module( 'filters', [] );
+
+momentFilter( filters );
+
+export default filters.name;
+
+
+
+
+
+
+//export default function( ngModule ) {
+//    momentFilter( ngModule );
+//}
