+++ app/src/oldMaterial/appOLDVERSION.html
@@ -1,0 +1,71 @@
+<div ng-controller='AppCtrl'>
+    <h1>Create New Magazine</h1>
+
+    <p>{{ magazine | json }}</p>
+
+
+    <magazine-edit magazine='magazine' saved='update(magazine)'></magazine-edit>
+
+
+
+</div>
+
+
+
+
+<hr>
+
+<div ng-controller='MagazineCtrl'>
+    <h1>List of Online News Portals</h1>
+    <div>
+        Set Search Name:    <input ng-model='search.name'><br>
+        Set Minimum Age: <input ng-model='search2.age'><br>
+        {{ search2 | json}}<br>
+    </div>
+
+    <table>
+        <thead>
+            <td>index</td>
+            <td>name</td>
+            <td>age</td>
+            <td>comment</td>
+        </thead>
+
+        <div>{{ avg }} is average age.</div>
+
+        <br>
+
+        <button ng-click="applyAverage()" >
+            {{ ageInput ? 'Show All' : 'Show Above Average Only' }} </button>
+
+        <br>
+        <tr ng-repeat="magazine in magazines
+                    | filter: search
+                    | filter: ageInput : biggerThanAverage
+                    | orderBy: 'age' ">
+
+
+            <td>{{ $index + 1 }}        </td>
+            <td>{{ magazine.name }}     </td>
+            <td>{{ magazine.age }}      </td>
+            <td>{{ magazine.comment }}  </td>
+            <td>{{ magazine.url }}  </td>
+            <td>{{ magazine.change }}  </td>
+
+
+        </tr>
+    </table>
+</div>
+
+
+<div ng-controller='AppCtrl'>
+    <h1>My app</h1>
+    <!--<p><input ng-model='format' value='MM YYYY'></p>-->
+    <h5>created {{ created | moment }}<h5>
+</div>
+
+<hr>
+
+
+
+
